corridor maze
